Creating setupbitrategraph util function and replacing everywhere it was called
